modularise the audio (also reimplement naudio and ditch managedbass bc managedbass sucks shit)

+++ BMSPlayer/BMS/ChannelEvent.cs
@@ -18,7 +18,7 @@ namespace Supernova.BMS
                 // We'll worry about #METER... later. Not now.
             }
         }
-        public int Measure = 0;
+        public float Measure = 0;
         public string Event = "00";
         public string Channel = "00";
     }
